+++ src/utils/index.js
@@ -1,6 +1,7 @@
 export { default as cloneSprite } from './clone-sprite';
 export { default as colorToDecAlpha } from './color-to-dec-alpha';
 export { default as CustomBufferResource } from './custom-buffer-resource';
+export { default as deserializeState } from './deserialize-state';
 export { default as ifNotNull } from './if-not-null';
 export { default as isNdarray } from './is-ndarray';
 export { default as getBBox } from './get-bbox';
@@ -11,6 +12,7 @@ export { default as createScale } from './scale';
 export { default as toAlignment } from './to-alignment';
 export { default as toDisplayObject } from './to-display-object';
 export { default as toHomogeneous } from './to-homogeneous';
+export { default as serializeState } from './serialize-state';
 export { default as uniqueStr } from './unique-str';
 export { default as whichTransitionEvent } from './which-transition-event';
 export { default as zoomToScale } from './zoom-to-scale';