+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- Add state serialization options via `exportState({ serialize: true })`
+
 _[Changes since v0.7.7](https://github.com/flekschas/piling.js/compare/v0.7.7...master)_
 
 ## v0.7.7